requirements.txt
# Hugging Face / model I/O
transformers==4.53.2
safetensors==0.5.3
tokenizers==0.21.2
huggingface-hub==0.34.3
hf-xet==1.1.5

# Media handling
decord==0.6.0
imageio==2.37.0
pdf2image==1.17.0
Pillow==11.0.0
streamlit

# Utilities
tqdm==4.67.1
python-dotenv==1.1.1
Jinja2==3.1.4
sympy==1.13.3
